reorganizing Util
