+++ src/nfa/tests-nfa.cc
@@ -2105,32 +2105,32 @@ TEST_CASE("Mata::Nfa::revert()")
 } // }}}
 
 // These are speed test of variants of revert, they take some time though.
-// TEST_CASE("Mata::Nfa::fragile_revert() speed") {
-//     Nfa B;
-//     FILL_WITH_AUT_A(B);
-//     //FILL_WITH_AUT_B(B);
-//     //FILL_WITH_AUT_C(B);
-//     //FILL_WITH_AUT_D(B);
-//     for (int i=0;i<300000;i++) {
-//         B = fragile_revert(B);
-//     }
-// }
-//
-// TEST_CASE("Mata::Nfa::simple_revert() speed") {
-//     Nfa B;
-//     FILL_WITH_AUT_A(B);
-//     for (int i=0;i<300000;i++) {
-//         B = simple_revert(B);
-//     }
-// }
-//
-// TEST_CASE("Mata::Nfa::somewhat_simple_revert() speed") {
-//     Nfa B;
-//     FILL_WITH_AUT_C(B);
-//     for (int i=0;i<300000;i++) {
-//         B = somewhat_simple_revert(B);
-//     }
-// }
+TEST_CASE("Mata::Nfa::fragile_revert() speed", "[.profiling],[revert]") {
+    Nfa B;
+    FILL_WITH_AUT_A(B);
+    //FILL_WITH_AUT_B(B);
+    //FILL_WITH_AUT_C(B);
+    //FILL_WITH_AUT_D(B);
+    for (int i=0;i<300000;i++) {
+        B = fragile_revert(B);
+    }
+}
+
+TEST_CASE("Mata::Nfa::simple_revert() speed", "[.profiling],[revert]" ) {
+    Nfa B;
+    FILL_WITH_AUT_A(B);
+    for (int i=0;i<300000;i++) {
+        B = simple_revert(B);
+    }
+}
+
+TEST_CASE("Mata::Nfa::somewhat_simple_revert() speed", "[.profiling],[revert]") {
+    Nfa B;
+    FILL_WITH_AUT_C(B);
+    for (int i=0;i<300000;i++) {
+        B = somewhat_simple_revert(B);
+    }
+}
 
 TEST_CASE("Mata::Nfa::is_deterministic()")
 { // {{{
